[documentation] Improve asymptote, build system
I am using a more verbose git commit message to please Yu-Ting. Hopefully this message and description better elucidates the intent and quality of my actions.

I updated the documentation on the build system to reflect current changes. In particular, the --dont-open flag was replaced by the --open flag, requiring opting in. Haha, "opt". What other funny three-letter words are there? Just considering three letter *animals*, there are already a lot: the ant, the dog, the cat, the bee, the cow; the YAK, the ELK, the EMU, the GNU. And yet, there is only one two-letter animal: the OX. Oh, the veritable ox. We just left the Chinese Year of the Ox (牛), although I don't know why that character—which, as far as I know, is generic for "cow"—specifically refers to oxen. What I do know is that us North Americans are superior English speakers compared to the rest of the world, because BrE, AuE, etc. speakers do not say "ox"; they say "steer" or "bullock". What joyless lives they must lead, devoid of two-letter animals!

Editing GAtM as an adult makes me feel like a gruff Tibetan yak... or, upon further consideration... like an ox. I feel hard-working, persistent—hooked up to the yoke with my ox brethren (Yu-Ting, mainly), plowing the Herreshoff-ian soil so that it may be fertile ground for student learning. Ox gang unite ✌️.

Anyway, back to my changes. I also added a link to an excellent document that describes most of the Asymptote language. I reference it to this day and it is unparalleled in ease of reading.